+++ Source/DREAM3DLib/ReconstructionFilters/AlignSections.cpp
@@ -197,15 +197,15 @@ void AlignSections::execute()
     std::stringstream ss;
     ss << "Aligning Sections - Transferring Cell Data - " << ((float)i/dims[2])*100 << " Percent Complete";
    // notifyProgress(ss.str(), 0, Observable::UpdateProgressMessage);
-    slice = (dims[2] - 1) - i;
+    slice = static_cast<int>( (dims[2] - 1) - i );
     for (DimType l = 0; l < dims[1]; l++)
     {
       for (DimType n = 0; n < dims[0]; n++)
       {
-        if(yshifts[i] >= 0) yspot = l;
-        if(xshifts[i] >= 0) xspot = n;
-        if(yshifts[i] < 0) yspot = dims[1] - 1 - l;
-        if(xshifts[i] < 0) xspot = dims[0] - 1 - n;
+        if(yshifts[i] >= 0) yspot = static_cast<int>(l);
+        if(xshifts[i] >= 0) xspot = static_cast<int>(n);
+        if(yshifts[i] < 0) yspot = static_cast<int>( dims[1] - 1 - l );
+        if(xshifts[i] < 0) xspot = static_cast<int>( dims[0] - 1 - n );
         newPosition = (slice * dims[0] * dims[1]) + (yspot * dims[0]) + xspot;
         currentPosition = (slice * dims[0] * dims[1]) + ((yspot + yshifts[i]) * dims[0]) + (xspot + xshifts[i]);
         if((yspot + yshifts[i]) >= 0 && (yspot + yshifts[i]) <= dims[1] - 1 && (xspot + xshifts[i]) >= 0
