+++ Source/DREAM3DLib/StatisticsFilters/FindAxisODF.cpp
@@ -195,7 +195,7 @@ void FindAxisODF::execute()
       OrientationMath::eulertoRod(r1, r2, r3, ea1, ea2, ea3);
       m_OrientationOps[Ebsd::CrystalStructure::OrthoRhombic]->getODFFZRod(r1, r2, r3);
       bin = m_OrientationOps[Ebsd::CrystalStructure::OrthoRhombic]->getOdfBin(r1, r2, r3);
-	  axisodf[m_FieldPhases[i]]->SetValue(bin, (axisodf[m_FieldPhases[i]]->GetValue(bin) + (1.0/totalaxes[m_FieldPhases[i]])));
+	  axisodf[m_FieldPhases[i]]->SetValue(bin, (axisodf[m_FieldPhases[i]]->GetValue(bin) + static_cast<float>( (1.0/totalaxes[m_FieldPhases[i]]) )));
     }
   }
  // int err;
